user teams util
